# backend/app/openapi/v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Expense Tracker V2 API",
    "version": "2.0.0",
    "description": "Authenticated transaction history and record management, transient assisted-entry drafts, and stateless Companion answers and reviewable transaction proposals, plus public health and fixed category endpoints."
  },
  "servers": [{ "url": "/api/v2" }],
  "paths": {
    "/companion/query": {
      "post": {
        "operationId": "queryCompanion",
        "summary": "Ask about transaction history or prepare one transaction proposal",
        "description": "Stateless. Without target_transaction_id, the operation returns only a read-only answer or clarification. With one selected in-scope transaction ID, the provider may prepare one validated edit, soft-delete, or expense-context proposal. A proposal_review request revalidates user-edited transient values and refreshes its deterministic effect without invoking the provider. Proposals are transient and do not mutate records; users confirm them through the existing PUT or DELETE /transactions/{id} command. Current month is determined by the backend in Europe/Madrid. The provider receives only the question and minimal deterministic evidence for the selected intent. Questions, answers, proposals, and evidence are not persisted or logged.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanionQueryRequest" } } }
        },
        "responses": {
          "200": { "description": "Validated transient answer, clarification, or unconfirmed transaction proposal and deterministic evidence.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanionQueryResponse" } } } },
          "400": { "description": "Question is blank, longer than 1,000 characters, scope or target is invalid, or edited proposal fields fail validation.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "409": { "description": "The selected record changed while its transient proposal was being edited." },
          "401": { "description": "Authentication is required." },
          "429": { "description": "The user's AI request limit has been reached.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "502": { "description": "The provider returned malformed or unsupported output.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "503": { "description": "The provider is unavailable or safely configured AI is not available.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "504": { "description": "The provider request timed out.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } }
        }
      }
    },
    "/assisted-entry/text": {
      "post": {
        "operationId": "createTextTransactionDraft",
        "summary": "Create a transient transaction draft from text",
        "description": "Uses only the submitted text, the server's current Europe/Madrid date, and fixed categories. Does not send transaction history, persist a draft, or create a transaction. The returned fields must be reviewed and saved through POST /transactions.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TextDraftRequest" } } }
        },
        "responses": {
          "200": { "description": "Transient editable draft and validation notes.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TextDraftResponse" } } } },
          "400": { "description": "Text is blank, over 1,000 characters, or malformed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Authentication is required." },
          "429": { "description": "The user's AI request limit has been reached.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "502": { "description": "The provider returned malformed or unusable output.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "503": { "description": "The provider is unavailable or safely configured AI is not available.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "504": { "description": "The provider request timed out.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } }
        }
      }
    },
    "/assisted-entry/receipt": {
      "post": {
        "operationId": "createReceiptTransactionDraft",
        "summary": "Create a transient transaction draft from one receipt image",
        "description": "Accepts one JPEG, PNG, or WebP image no larger than 10 MB. The backend validates file count, declared media type, size, and image decodability before starting provider processing. The provider receives only the image, its media type, the server's current Europe/Madrid date, and fixed expense categories. The upload, extracted content, and draft are transient and are not stored or logged. Review and save any values through POST /transactions.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/ReceiptDraftRequest" },
              "encoding": { "receipt": { "contentType": "image/jpeg, image/png, image/webp" } }
            }
          }
        },
        "responses": {
          "200": { "description": "Transient editable expense draft and validation notes.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReceiptDraftResponse" } } } },
          "400": { "description": "The upload is missing, contains multiple files, uses an unsupported media type, or is invalid or damaged.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Authentication is required." },
          "413": { "description": "The image exceeds the 10 MB upload limit.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "429": { "description": "The user's AI request limit has been reached.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "502": { "description": "The provider returned malformed or unusable output.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "503": { "description": "The provider is unavailable or safely configured AI is not available.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } },
          "504": { "description": "The provider request timed out.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AIErrorResponse" } } } }
        }
      }
    },
    "/transactions": {
      "get": {
        "operationId": "listTransactions",
        "summary": "List the authenticated user's transactions",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "scope", "in": "query", "schema": { "type": "string", "enum": ["current_month", "all"], "default": "current_month" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 } }
        ],
        "responses": {
          "200": { "description": "Current page and scope-wide summary.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionListResponse" } } } },
          "400": { "description": "Invalid query parameters.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Authentication is required." }
        }
      },
      "post": {
        "operationId": "createTransaction",
        "summary": "Create an income or expense transaction",
        "description": "Creates one transaction for the authenticated user. Expense requests may include optional b_u_c and reflective_context enum values or null; income requests must omit both fields. A repeated Idempotency-Key within 24 hours returns the original successful response.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "Idempotency-Key",
          "in": "header",
          "required": true,
          "description": "A fresh UUID v4 for each new submission.",
          "schema": { "type": "string", "format": "uuid" }
        }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTransactionRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Transaction created, or original success returned for a repeated key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTransactionResponse" } } }
          },
          "400": {
            "description": "Invalid transaction fields or idempotency key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "401": { "description": "Authentication is required." }
        }
      }
    },
    "/transactions/{transaction_id}": {
      "parameters": [{ "name": "transaction_id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }],
      "get": {
        "operationId": "getTransaction",
        "summary": "Retrieve one current transaction for editing",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "The transaction.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionResponse" } } } },
          "401": { "description": "Authentication is required." },
          "403": { "description": "The transaction belongs to another user." },
          "404": { "description": "The transaction is missing or soft-deleted." }
        }
      },
      "put": {
        "operationId": "updateTransaction",
        "summary": "Update an income or expense transaction",
        "description": "Transaction type is immutable. The submitted updated_at value must match the current record. For expenses, context fields may be omitted to preserve their current values or set to null to clear them. Income requests must omit context fields.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTransactionRequest" } } } },
        "responses": {
          "200": { "description": "Updated transaction and affected month range.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTransactionResponse" } } } },
          "400": { "description": "Invalid fields or idempotency key.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Authentication is required." },
          "403": { "description": "The transaction belongs to another user." },
          "404": { "description": "The transaction is missing or soft-deleted." },
          "409": { "description": "The supplied updated_at value is stale." }
        }
      },
      "delete": {
        "operationId": "deleteTransaction",
        "summary": "Soft-delete a transaction",
        "description": "Sets deleted_at and returns the deleted identifier. Repeating the same key within 24 hours returns the original response.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Deleted record identifier and affected period.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteTransactionResponse" } } } },
          "400": { "description": "Invalid idempotency key.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Authentication is required." },
          "403": { "description": "The transaction belongs to another user." },
          "404": { "description": "The transaction is missing or soft-deleted." }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "CompanionQueryRequest": {
        "type": "object",
        "required": ["question", "scope"],
        "properties": {
          "question": { "type": "string", "minLength": 1, "maxLength": 1000 },
          "scope": { "type": "string", "enum": ["current_month", "all"] },
          "target_transaction_id": { "type": "integer", "minimum": 1, "description": "One current transaction selected from the bounded matching-record result. Proposals are allowed only for this owned, non-deleted, in-scope record." },
          "proposal_review": { "$ref": "#/components/schemas/CompanionProposalReview" }
        },
        "additionalProperties": false
      },
      "CompanionQueryResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "oneOf": [
              { "$ref": "#/components/schemas/CompanionAnswer" },
              { "$ref": "#/components/schemas/CompanionClarification" },
              { "$ref": "#/components/schemas/CompanionProposalResponse" }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "TextDraftRequest": {
        "type": "object",
        "required": ["text"],
        "properties": { "text": { "type": "string", "minLength": 1, "maxLength": 1000, "example": "I spent €12.50 on lunch today" } },
        "additionalProperties": false
      },
...
      "TextDraftResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["draft", "missing_fields", "uncertainties"],
            "properties": {
              "draft": { "$ref": "#/components/schemas/TextDraft" },
              "missing_fields": { "type": "array", "items": { "type": "string", "enum": ["transaction_type", "amount", "transaction_date", "category_key"] } },
              "uncertainties": { "type": "array", "items": { "$ref": "#/components/schemas/TextDraftUncertainty" } }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "ReceiptDraftRequest": {
        "type": "object",
        "required": ["receipt"],
        "properties": {
          "receipt": { "type": "string", "format": "binary", "description": "Exactly one JPEG, PNG, or WebP image, at most 10 MB." }
        },
        "additionalProperties": false
      },
...
      "ReceiptDraftResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["draft", "missing_fields", "uncertainties"],
            "properties": {
              "draft": { "$ref": "#/components/schemas/ReceiptDraft" },
              "missing_fields": { "type": "array", "items": { "type": "string", "enum": ["transaction_type", "amount", "transaction_date", "category_key"] } },
              "uncertainties": { "type": "array", "items": { "$ref": "#/components/schemas/TextDraftUncertainty" } }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "AIErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["message", "code"],
            "properties": { "message": { "type": "string" }, "code": { "type": "string" } }
          }
        }
      },
...
      "CreateTransactionResponse": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": { "$ref": "#/components/schemas/Transaction" },
          "meta": {
            "type": "object",
            "required": ["affected_period_range"],
            "properties": { "affected_period_range": { "$ref": "#/components/schemas/AffectedPeriodRange" } },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "TransactionResponse": {
        "type": "object", "required": ["data"],
        "properties": { "data": { "$ref": "#/components/schemas/Transaction" } },
        "additionalProperties": false
      },
...
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["message", "code"],
            "properties": {
              "message": { "type": "string" },
              "code": { "type": "string", "enum": ["VALIDATION_ERROR", "IDEMPOTENCY_KEY_REQUIRED"] },
              "fields": { "type": "object", "additionalProperties": { "type": "string" } }
            }
          }
        }
      },
...
    }
  }
}